refactor: split frontend/backend god files into modules
Refs #42
