+++ src-tauri/src/app/mod.rs
@@ -1,0 +1,3 @@
+pub mod commands;
+pub mod shutdown;
+pub mod state;
