+++ src/shared/i18n/locales.ts
@@ -1,0 +1,313 @@
+import type { Language, LocaleTexts } from "../contracts";
+
+export const LOCALES: Record<Language, LocaleTexts> = {
+  es: {
+    headerTitle: "Android Emulator Proxy Console",
+    headerSubhead:
+      "Trazado HTTP/HTTPS local para emuladores Android, con captura, filtros, detalle e interceptación editable.",
+    settings: "Configuracion",
+    settingsTitle: "Preferencias",
+    language: "Idioma",
+    theme: "Tema",
+    fontSize: "Tamano de fuente",
+    sensitiveData: "Datos sensibles",
+    certTrusted: "CA confiada en emulador",
+    closeSettings: "Cerrar configuracion",
+    light: "Claro",
+    dark: "Oscuro",
+    fontSmall: "Chico",
+    fontMedium: "Medio",
+    fontLarge: "Grande",
+    showSensitive: "Mostrar valores reales",
+    hideSensitive: "Ocultar valores sensibles",
+    markCertTrusted: "Confiada",
+    markCertPending: "Pendiente",
+    controls: "Controles de tracing",
+    emulator: "Emulador",
+    proxyHost: "Proxy host (emulador)",
+    proxyPort: "Proxy port",
+    refresh: "Actualizar",
+    prepareCa: "Preparar instalacion de CA",
+    certInstallConsentTitle: "Permisos para instalacion de certificado",
+    certInstallConsentBody:
+      "La app copiara el certificado al emulador e intentara abrir Security para que completes la instalacion manual.",
+    certInstallFlowLabel: "Flujo recomendado",
+    certInstallFlowDesc:
+      "1) Copiar certificado en Download. 2) Abrir Security. 3) En Android: Encryption & credentials > Install a certificate > CA certificate.",
+    certInstallContinue: "Continue",
+    certInstallPreparing: "Copiando certificado y abriendo Security...",
+    exitPromptTitle: "Antes de salir",
+    exitPromptBody:
+      "Vamos a cerrar la app y limpiar el proxy del emulador para evitar que quede sin internet.",
+    exitPromptProxyHint:
+      "Si cierras forzado y el proxy no se limpia, ejecuta: adb shell settings put global http_proxy :0",
+    exitPromptCertHint:
+      "El certificado CA se elimina manualmente en Android: Security > Encryption & credentials.",
+    exitPromptCancel: "Cancelar",
+    exitPromptConfirm: "Salir y limpiar",
+    exitPromptClosing: "Cerrando app y limpiando proxy...",
+    startTracing: "Iniciar trazado",
+    stopTracing: "Detener trazado",
+    operationStatus: "Estados operativos",
+    adbMissing: "ADB missing",
+    noEmulator: "No emulator",
+    certNotTrusted: "Cert not trusted",
+    tracingActive: "Tracing active",
+    ready: "Ready",
+    adbMissingDesc: "No se encontró adb local para controlar el emulador.",
+    noEmulatorDesc: "ADB responde, pero no hay emuladores online.",
+    certNotTrustedDesc: "La CA local aun no fue marcada como confiada para HTTPS.",
+    tracingActiveDesc: "El proxy MITM esta activo y capturando requests.",
+    readyDesc: "ADB y emulador listos para iniciar trazado.",
+    adbMissingAction: "Instala platform-tools o define ADB_PATH.",
+    noEmulatorAction: "Inicia un AVD y verifica `adb devices`.",
+    certNotTrustedAction: "Instala la CA en el emulador y marca como confiada.",
+    tracingActiveAction: "Puedes inspeccionar, filtrar o interceptar requests.",
+    readyAction: "Presiona Start Tracing para comenzar.",
+    adbPanel: "ADB & Emuladores",
+    sessionPanel: "Sesion de tracing",
+    adbAvailable: "ADB disponible",
+    adbVersion: "Version ADB",
+    adbPath: "Ruta ADB",
+    emulatorsConnected: "Emuladores conectados",
+    state: "Estado",
+    activeEmulator: "Emulador activo",
+    proxyApplied: "Proxy aplicado",
+    localCa: "CA local",
+    startedAt: "Iniciado en",
+    active: "Activo",
+    stopped: "Detenido",
+    requestsTitle: "Requests capturadas",
+    workspaceTabRequests: "Requests",
+    workspaceTabInterception: "Intercepcion",
+    clearSession: "Clear Session",
+    filterText: "Buscar por host o path",
+    filterMethod: "Filtro metodo HTTP",
+    filterStatus: "Status: 200, 2xx, 400-499",
+    allMethods: "Todos los metodos",
+    clearFilters: "Limpiar filtros",
+    sortBy: "Ordenar por",
+    sortDirection: "Direccion",
+    sortById: "ID",
+    sortByTime: "Tiempo",
+    asc: "ASC",
+    desc: "DESC",
+    timestamp: "Timestamp",
+    method: "Method",
+    host: "Host",
+    path: "Path",
+    status: "Status",
+    duration: "Duracion",
+    noTraffic: "Sin trafico capturado aun.",
+    noFilterResults: "No hay resultados para los filtros actuales.",
+    detailTitle: "Detalle de request",
+    selectRequest: "Selecciona una request para ver request, response y metadata.",
+    detailsFor: "Detalle",
+    copied: "Copiado al portapapeles.",
+    copyFailed: "No fue posible copiar al portapapeles.",
+    copyUrl: "Copiar URL",
+    exportCurl: "Exportar como cURL",
+    requestTab: "Request",
+    responseTab: "Response",
+    headersTab: "Headers",
+    cookiesTab: "Cookies",
+    paramsTab: "Params",
+    timingTab: "Timing",
+    requestSize: "Request size",
+    responseSize: "Response size",
+    noBody: "No hay body textual disponible para esta captura.",
+    copyContent: "Copiar contenido",
+    requestHeaders: "Request headers",
+    responseHeaders: "Response headers",
+    noHeaders: "Sin headers",
+    noCookies: "No se detectaron cookies en request/response.",
+    noParams: "No se detectaron query params para esta request.",
+    queryLabel: "Query",
+    requestBodySize: "Request body size",
+    responseBodySize: "Response body size",
+    interceptStatus: "Estado interceptacion",
+    originalRequest: "Original",
+    interceptionTitle: "Intercepcion y reenvio",
+    interceptionEnabled: "Modo interceptacion",
+    interceptionTimeout: "Timeout (ms)",
+    interceptionHostFilter: "Dominio contiene",
+    interceptionPathFilter: "Path contiene",
+    interceptionMethodFilter: "Metodo",
+    interceptionAllMethods: "Todos",
+    interceptionRuleEnabled: "Activa",
+    interceptionRules: "Reglas",
+    interceptionRulesEmptyHint: "Sin reglas activas: se intercepta todo.",
+    manageRules: "Reglas",
+    close: "Cerrar",
+    addRule: "Agregar regla",
+    removeRule: "Quitar",
+    applyInterception: "Aplicar",
+    pendingInterceptions: "Pendientes",
+    noPendingInterceptions: "No hay requests pendientes de decision.",
+    interceptEditor: "Editor de request interceptada",
+    editorHeaders: "Headers (uno por linea: Nombre: Valor)",
+    editorBody: "Body",
+    editorQuery: "Query string",
+    editorCookies: "Cookies",
+    forwardRequest: "Reenviar",
+    dropRequest: "Descartar",
+    intercepted: "interceptada",
+    timeout: "timeout",
+    dropped: "descartada",
+  },
+  en: {
+    headerTitle: "Android Emulator Proxy Console",
+    headerSubhead:
+      "Local HTTP/HTTPS tracing for Android emulators, with capture, filters, details, and editable interception.",
+    settings: "Settings",
+    settingsTitle: "Preferences",
+    language: "Language",
+    theme: "Theme",
+    fontSize: "Font size",
+    sensitiveData: "Sensitive data",
+    certTrusted: "CA trusted on emulator",
+    closeSettings: "Close settings",
+    light: "Light",
+    dark: "Dark",
+    fontSmall: "Small",
+    fontMedium: "Medium",
+    fontLarge: "Large",
+    showSensitive: "Show real values",
+    hideSensitive: "Mask sensitive values",
+    markCertTrusted: "Trusted",
+    markCertPending: "Pending",
+    controls: "Tracing controls",
+    emulator: "Emulator",
+    proxyHost: "Proxy host (emulator)",
+    proxyPort: "Proxy port",
+    refresh: "Refresh",
+    prepareCa: "Prepare CA Install",
+    certInstallConsentTitle: "Certificate install permissions",
+    certInstallConsentBody:
+      "The app will copy the certificate to the emulator and attempt to open Security so you can complete manual install.",
+    certInstallFlowLabel: "Recommended flow",
+    certInstallFlowDesc:
+      "1) Copy certificate into Download. 2) Open Security. 3) In Android: Encryption & credentials > Install a certificate > CA certificate.",
+    certInstallContinue: "Continue",
+    certInstallPreparing: "Copying certificate and opening Security...",
+    exitPromptTitle: "Before exit",
+    exitPromptBody:
+      "The app will close and clear emulator proxy settings so the device keeps normal connectivity.",
+    exitPromptProxyHint:
+      "If the app is force-closed and proxy remains, run: adb shell settings put global http_proxy :0",
+    exitPromptCertHint:
+      "CA certificate removal is manual in Android: Security > Encryption & credentials.",
+    exitPromptCancel: "Cancel",
+    exitPromptConfirm: "Exit and clean",
+    exitPromptClosing: "Closing app and cleaning proxy...",
+    startTracing: "Start Tracing",
+    stopTracing: "Stop Tracing",
+    operationStatus: "Operational states",
+    adbMissing: "ADB missing",
+    noEmulator: "No emulator",
+    certNotTrusted: "Cert not trusted",
+    tracingActive: "Tracing active",
+    ready: "Ready",
+    adbMissingDesc: "adb binary was not found on this machine.",
+    noEmulatorDesc: "ADB is reachable but no online emulator was found.",
+    certNotTrustedDesc: "Local CA is not marked as trusted for HTTPS yet.",
+    tracingActiveDesc: "MITM proxy is active and capturing traffic.",
+    readyDesc: "ADB and emulator are ready to start tracing.",
+    adbMissingAction: "Install platform-tools or set ADB_PATH.",
+    noEmulatorAction: "Start an AVD and verify `adb devices`.",
+    certNotTrustedAction: "Install CA in emulator and mark as trusted.",
+    tracingActiveAction: "You can inspect, filter, or intercept requests.",
+    readyAction: "Press Start Tracing to begin.",
+    adbPanel: "ADB & Emulators",
+    sessionPanel: "Tracing session",
+    adbAvailable: "ADB available",
+    adbVersion: "ADB version",
+    adbPath: "ADB path",
+    emulatorsConnected: "Connected emulators",
+    state: "State",
+    activeEmulator: "Active emulator",
+    proxyApplied: "Applied proxy",
+    localCa: "Local CA",
+    startedAt: "Started at",
+    active: "Active",
+    stopped: "Stopped",
+    requestsTitle: "Captured requests",
+    workspaceTabRequests: "Requests",
+    workspaceTabInterception: "Interception",
+    clearSession: "Clear Session",
+    filterText: "Search by host or path",
+    filterMethod: "HTTP method filter",
+    filterStatus: "Status: 200, 2xx, 400-499",
+    allMethods: "All methods",
+    clearFilters: "Clear filters",
+    sortBy: "Sort by",
+    sortDirection: "Direction",
+    sortById: "ID",
+    sortByTime: "Time",
+    asc: "ASC",
+    desc: "DESC",
+    timestamp: "Timestamp",
+    method: "Method",
+    host: "Host",
+    path: "Path",
+    status: "Status",
+    duration: "Duration",
+    noTraffic: "No traffic captured yet.",
+    noFilterResults: "No results for the current filters.",
+    detailTitle: "Request details",
+    selectRequest: "Select a request to view request, response, and metadata.",
+    detailsFor: "Details",
+    copied: "Copied to clipboard.",
+    copyFailed: "Unable to copy to clipboard.",
+    copyUrl: "Copy URL",
+    exportCurl: "Export as cURL",
+    requestTab: "Request",
+    responseTab: "Response",
+    headersTab: "Headers",
+    cookiesTab: "Cookies",
+    paramsTab: "Params",
+    timingTab: "Timing",
+    requestSize: "Request size",
+    responseSize: "Response size",
+    noBody: "No textual body available for this capture.",
+    copyContent: "Copy content",
+    requestHeaders: "Request headers",
+    responseHeaders: "Response headers",
+    noHeaders: "No headers",
+    noCookies: "No cookies detected in request/response.",
+    noParams: "No query params detected for this request.",
+    queryLabel: "Query",
+    requestBodySize: "Request body size",
+    responseBodySize: "Response body size",
+    interceptStatus: "Interception status",
+    originalRequest: "Original",
+    interceptionTitle: "Interception & replay",
+    interceptionEnabled: "Interception mode",
+    interceptionTimeout: "Timeout (ms)",
+    interceptionHostFilter: "Host contains",
+    interceptionPathFilter: "Path contains",
+    interceptionMethodFilter: "Method",
+    interceptionAllMethods: "All",
+    interceptionRuleEnabled: "Enabled",
+    interceptionRules: "Rules",
+    interceptionRulesEmptyHint: "No active rules: intercept everything.",
+    manageRules: "Rules",
+    close: "Close",
+    addRule: "Add rule",
+    removeRule: "Remove",
+    applyInterception: "Apply",
+    pendingInterceptions: "Pending",
+    noPendingInterceptions: "No requests pending decision.",
+    interceptEditor: "Intercepted request editor",
+    editorHeaders: "Headers (one per line: Name: Value)",
+    editorBody: "Body",
+    editorQuery: "Query string",
+    editorCookies: "Cookies",
+    forwardRequest: "Forward",
+    dropRequest: "Drop",
+    intercepted: "intercepted",
+    timeout: "timeout",
+    dropped: "dropped",
+  },
+};
+
