+++ src/shared/utils/clipboard.ts
@@ -1,0 +1,23 @@
+export async function copyToClipboard(content: string): Promise<void> {
+  if (!content) {
+    throw new Error("Empty content");
+  }
+
+  if (navigator?.clipboard?.writeText) {
+    await navigator.clipboard.writeText(content);
+    return;
+  }
+
+  const textarea = document.createElement("textarea");
+  textarea.value = content;
+  textarea.setAttribute("readonly", "");
+  textarea.style.position = "absolute";
+  textarea.style.left = "-9999px";
+  document.body.appendChild(textarea);
+  textarea.select();
+  const ok = document.execCommand("copy");
+  document.body.removeChild(textarea);
+  if (!ok) {
+    throw new Error("copy command failed");
+  }
+}
